displaying the data on the page

+++ menu/models.py
@@ -11,7 +11,7 @@ class Catigory(models.Model):
 class MenuItem(models.Model):
     name = models.CharField(max_length=50 , default='')
     description = models.TextField(default='')
-    price = models.IntegerField(default=0)
+    price = models.FloatField(default=0)
     photo = models.ImageField(default='' , upload_to='testImages/%y/%m/%d')
     restaurant = models.ForeignKey(Restaurant , on_delete=models.CASCADE , default='')
     catigory = models.ForeignKey(Catigory , on_delete=models.CASCADE , default='')
